# Fixes #36943 - reduce the amount of queries to get the errata counts