Loading the all hosts page (http://myforeman/hosts) with about 72 hosts.

```
OLD
Completed 200 OK in 9900ms (Views: 8613.5ms | ActiveRecord: 787.6ms | Allocations: 5520271)
Completed 200 OK in 9595ms (Views: 8381.8ms | ActiveRecord: 789.9ms | Allocations: 5524887)
SQL Queries: 195

irb(main):049:0> total=0; 100.times do x=Benchmark.measure { Host.all.each do |h| h.content_facet.nil? || h.content_facet.errata_counts_old end }; total = total + x.total; sleep 0.5; end; puts total;
68.09449500000002

NEW
Completed 200 OK in 10789ms (Views: 9794.5ms | ActiveRecord: 549.5ms | Allocations: 5511210)
Completed 200 OK in 9244ms (Views: 8400.2ms | ActiveRecord: 385.6ms | Allocations: 5511208)
SQL Queries: 65

irb(main):050:0> total=0; 100.times do x=Benchmark.measure { Host.all.each do |h| h.content_facet.nil? || h.content_facet.errata_counts end }; total = total + x.total; sleep 0.5; end; puts total;
37.18352599999951
```