@sbernhard 2023-12-22
foreman rake benchmark for about 80 hosts.

Old
```
irb(main):003:0> total=0; 100.times do x=Benchmark.measure { Host.all.each do |h| h.content_facet.nil? || h.content_facet.errata_counts_old; end }; total = total + x.total; end
=> 100
irb(main):004:0> puts total
37.80042699999998
```

New:
```
irb(main):001:0> total=0; 100.times do x=Benchmark.measure { Host.all.each do |h| h.content_facet.nil? || h.content_facet.errata_counts; end }; total = total + x.total; end
=> 100
irb(main):002:0> puts total
22.924028999999997
```